requirements.txt
python-telegram-bot[job-queue]==20.5
python-dotenv==1.0.0
pillow==10.0.1